Backend separation is underway.
DataFrame concept types are working as expected. Did implemenation for Arraymancer backend, needs cleanup to make the types less generic (could match other backend storage types if they existed). Added two new debug and testing utilities: "stringify" which does what a C programmer thinks it does, and "trace_result", which takes a single untyped argument, stringifies the node, and then places it as an argument in an echo command, along with it's stringified version as a lable. This is really helpful for quickly printing off the values of expressions, since the output is labelled with the expression that made it.